add docker-compose.yml
