+++ hw17-docker/Dockerfile
@@ -19,7 +19,6 @@ ENV PROJECT_DIR=/opt/project
 
 RUN mkdir -p $PROJECT_DIR
 WORKDIR $PROJECT_DIR
-COPY --from=0 $PROJECT_DIR/target/hw17-docker-2020-05-SNAPSHOT.jar $PROJECT_DIR/
+COPY --from=build $PROJECT_DIR/target/hw17-docker-2020-05-SNAPSHOT.jar $PROJECT_DIR/
 
-EXPOSE 8080
-CMD ["java","-jar","/opt/project/hw17-docker-2020-05-SNAPSHOT.jar"]
+ENTRYPOINT  ["java","-Dmongo.db.name=hw17Docker", "-Dmongo.host=mongo", "-Dmongo.port=27017","-jar","/opt/project/hw17-docker-2020-05-SNAPSHOT.jar"]
