+++ hw17-docker/docker-compose.yml
@@ -1,0 +1,35 @@
+version: "3"
+services:
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    privileged: true
+    ports:
+      - 8080:8080
+    environment:
+      - WAIT_HOSTS:mongo:27017
+    depends_on:
+      - mongo
+    links:
+      - mongo
+    networks:
+      - shared-net
+
+  mongo:
+    container_name: mongo
+    hostname: mongo
+    image: library/mongo:4.2.10
+    volumes:
+      - ./data:/data/db
+    ports:
+      - 27017:27017
+    networks:
+      - shared-net
+
+volumes:
+  mongo_data:
+
+networks:
+  shared-net:
+    driver: bridge
