+++ hw17-docker/src/main/java/ru/otus/hw17/config/MongoConfig.java
@@ -1,18 +1,13 @@
 package ru.otus.hw17.config;
 
-import com.mongodb.MongoClientSettings;
 import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoDatabase;
 import lombok.Data;
-import org.bson.codecs.configuration.CodecRegistry;
-import org.bson.codecs.pojo.PojoCodecProvider;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
+import org.springframework.data.mongodb.core.MongoClientFactoryBean;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Component;
-import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
-import static org.bson.codecs.configuration.CodecRegistries.fromRegistries;
 
 @Data
 @Component
@@ -21,11 +16,14 @@ public class MongoConfig {
 
   private int port;
   private String database;
-  private String uri;
+  private String host;
 
   @Bean
-  public MongoClient mongoClient() {
-    return MongoClients.create(createMongoClientSettings());
+  public MongoClientFactoryBean mongo() {
+    MongoClientFactoryBean mongo = new MongoClientFactoryBean();
+    mongo.setHost(host);
+    mongo.setPort(port);
+    return mongo;
   }
 
   @Bean
@@ -36,10 +34,5 @@ public class MongoConfig {
   @Bean
   public MongoTemplate mongoTemplate(MongoClient mongoClient) {
     return new MongoTemplate(mongoClient, database);
-  }
-
-  private MongoClientSettings createMongoClientSettings() {
-    CodecRegistry pojoCodecRegistry = fromRegistries(MongoClientSettings.getDefaultCodecRegistry(), fromProviders(PojoCodecProvider.builder().automatic(true).build()));
-    return MongoClientSettings.builder().codecRegistry(pojoCodecRegistry).build();
   }
 }
