+++ hw17-docker/src/test/java/ru/otus/hw17/repository/AuthorRepositoryTest.java
@@ -9,8 +9,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
-import ru.otus.hw17.config.MongoConfig;
 import ru.otus.hw17.model.Author;
+import ru.otus.hw17.test.config.MongoConfig;
 
 import java.util.Optional;
